Move PreservationEventSerializer to dor.providers.serializers, add test
