+++ dor/providers/process_basic_image.py
@@ -15,6 +15,7 @@ from dor.builders.parts import FileInfo, UseFunction, UseFormat, flatten_use
 from dor.providers.models import (
     Agent, AlternateIdentifier, FileReference, PackageResource, FileMetadata, PreservationEvent
 )
+from dor.providers.serializers import PreservationEventSerializer
 from dor.settings import template_env
 
 
@@ -23,30 +24,6 @@ ACCEPTED_IMAGE_MIMETYPES = [
     ImageMimetype.TIFF,
     ImageMimetype.JP2
 ]
-
-PREMIS_EVENT_TEMPLATE = template_env.get_template("premis_event.xml")
-
-
-class PreservationEventSerializer():
-
-    template = template_env.get_template("premis_event.xml")
-
-    def __init__(self, event: PreservationEvent):
-        self.event = event
-
-    def serialize(self) -> str:
-        event_data = {
-            "identifier": self.event.identifier,
-            "type": self.event.type,
-            "date_time": self.event.datetime.strftime("%Y-%m-%dT%H:%M:%SZ"),
-            "detail": self.event.detail,
-            "linking_agent": {
-                "type": self.event.agent.role,
-                "value": self.event.agent.address
-            }
-        }
-        return self.template.render(event=event_data)
-
 
 def create_preservation_event(
     type: str,
